requirements.txt
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
SQLAlchemy==1.3.23
Werkzeug==1.0.1
zope.interface==5.4.0
